Fix .data sections misidentified as .rodata by dtk
dtk can't always reliably identify REL sections in its initial analysis.
This is a manual fix - the list of RELs to fix was found by looking
at supposed .rodata splits that contained an fBase vtable, since vtables
should be in .data. This fix is required for scripted creation of REL actors
based on rel .data